requirements/base.txt
asgiref==3.5.0
Django==3.2.11
djangorestframework==3.13.1
psycopg2==2.9.3
pytz==2021.3
sqlparse==0.4.2
drf-yasg==1.20.0
django-dramatiq==0.10.0
redis==4.2.2